IGBF-3659

+++ core/ucsc-rest-api-service/src/main/java/org/lorainelab/igb/ucsc/rest/api/service/model/SchemaData.java
@@ -25,7 +25,7 @@ public class SchemaData {
     private String shortLabel;
     private String type;
     private String longLabel;
-    private int itemCount;
+    private long itemCount;
     private String parent;
     private String colorByStrand;
     private String visibility;
